add clear cmd
